+++ include/pmc/pmc_neigh_coloring.h
